Add dashboards rpm assemble support in opensearch-build repo (#1852)
* Add 1st draft of dashboards rpm bundle code


* Update test cases


* Remove unnecessary chown


* Remove changing /tmp permission due to PA in opensearch

+++ src/assemble_workflow/bundle_rpm.py
@@ -89,7 +89,10 @@ class BundleRpm:
         os.environ.pop('OPENSEARCH_PATH_CONF', None)
 
         # Restore config file and core folder to original location
-        shutil.move(f"{min_bin_env_path}.backup", min_bin_env_path)
+        if os.path.exists(f"{min_bin_env_path}.backup"):
+            logging.info(f"Restore {min_bin_env_path}.backup to {min_bin_env_path}")
+            shutil.move(f"{min_bin_env_path}.backup", min_bin_env_path)
+
         shutil.move(min_dest_path, min_source_path)
 
         # Run bundle rpmbuild
